Many changes and additions
Remove safe types in favor of better compiler warnings.
Add some helper constexpr functions, such as T max<T>().
Add function for specifically clearing flag bits (and update test records).
Fix bug in block, atomic, and nonatomic.
Make bv() only take an int.
Make getReg() return a reference.
Reorganise hierarchy.
Add heap type for careful dynamic memory allocation.
Remove Atomic class.
Make atomic(), block(), and nonatomic() more flexible.

+++ src/hardware/chip.hpp
@@ -28,11 +28,11 @@
 #ifndef NBOS_CHIP_HPP
 #define NBOS_CHIP_HPP
 
-#include "safe.hpp"
+#include "../primitive.hpp"
 
 namespace nbos {
 
-template <class T, Integer addr>
+template <class T, size_t addr>
 struct _reg {
     T*const p = (T*)(addr);
 };
